feat:create-bzero

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: masayama <masayama@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 11:43:18 by masayama          #+#    #+#             */
-/*   Updated: 2024/04/17 11:59:55 by masayama         ###   ########.fr       */
+/*   Updated: 2024/04/18 09:09:28 by masayama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	if (s == NULL)
